separate structure class from testentity
